backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
django-filter==23.5
pydub==0.25.1
mutagen==1.47.0